Bugfix page snaps to top
The window snapped to the top when clicking next slide because using href="#" will do that.
I would use javascript:void(0); but google really doesn't like that and sends many errors in console.
you could rewrite all the anchors into buttons, but that requires rewriting alot of code.
though there is still the problem of not being able to go back to the previous webpage in one click. now it takes 2 or 3.