Temp fix for Pool gallery
The pool gallery slideshow doesn't work so this is just to disable the slideshow whenever you are on the gallery.
The slideshow prevents clicking the cover page from going to its link so you would have to middle mouse over it which is annoying.
All you would see in the gallery is just all the cover pages anyways.